recognition/requirements.txt
tensorflow
tensorflow-datasets
pillow
numpy